refactor: extract rendering logic from Deck into DeckRenderer
Split runtime/deck.py by extracting all rendering pipelines (key,
touchscreen, info-screen) and related helpers into a new
DeckRenderer class in runtime/renderer.py.

- DeckRenderer owns: render_all_keys, render_dui_key,
  render_touchscreen, render_info_screen, drain_card_callbacks,
  is_dui_key, is_animating, _resolve_stylesheet, apply_theme
- Deck retains: lifecycle (start/stop), event dispatch, timeout
  polling, screen management, brightness, and thin delegation
  wrappers for backward compatibility
- No change to public Deck API
- deck.py reduced from 882 to 631 lines

Closes #195
